ssp: clean-up register setup
The SSP configuration is a total mess. This is a first pass at trying
to fix the setup of SSCR0..5, SSPSP, SFIFOTT

I went through each register one by one to figure out when it was
supposed to be programmed and with what value. The results seem to
improve a bit with a lot more repeatability and the bitclock/fs which
don't toggle when there is no audio.

I am far from done but since I'll be away from my desk for nearly two
weeks I could use some comments.

Known issues:
1. in I2S mode the data is not delayed by 1 block period despite setting DMYSTRT, something is off (this was already broken before the updates)
2. in 16 bit mode, the data is not well aligned on the 20 bit frame

Likely bugs (look for FIXMEs)
1. save/restore only deals with a couple of registers
2. why were SSCR1 PINTE and RWOT set?
3. what is the definition of SSP_CLK_DEFAULT
4. do we need to set SCFR when codec is master? The behavior could be codec
dependent, nothing says the FS needs to remain on outside transfers
5. not sure how LEFT_J is enabled
6. watermarks are likely incorrect

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>